Moved TestTokenizer and HardcodedSource for better separation of test and runtime classpaths. Modified Eclipse configuration to mark packages under test/ as containing test classes in order to prevent unwanted references from runtime packages to test packages.
